scripts/requirements.txt
requests>=2.31.0
gitpython>=3.1.40
python-dotenv>=1.0.0 